@@ -1,5 +1,12 @@
 module blazium.app
 
-go 1.22.4
+go 1.23
 
-require github.com/gorilla/mux v1.8.1 // indirect
+toolchain go1.23.2
+
+require (
+	github.com/PuerkitoBio/goquery v1.10.0 // indirect
+	github.com/andybalholm/cascadia v1.3.2 // indirect
+	github.com/gorilla/mux v1.8.1 // indirect
+	golang.org/x/net v0.29.0 // indirect
+)
